+++ src/test/java/divisionTest.java
@@ -5,8 +5,10 @@ import org.testng.annotations.Test;
 public class divisionTest {
     @Test
     public void divisionTest(){
-        double result = calculator.division(12,3);
-        double expected = 4;
+        Integer result = calculator.division(12,3);
+        Integer expected = 4;
         Assert.assertEquals(expected,result);
     }
+
+
 }
