+++ src/test/java/remainderTest.java
@@ -5,8 +5,8 @@ import org.testng.annotations.Test;
 public class remainderTest {
     @Test
     public void remainderTest(){
-        double result = calculator.remainder(7,2);
-        double expected = 1;
+        Integer result = calculator.remainder(7,2);
+        Integer expected = 1;
         Assert.assertEquals(expected,result);
     }
 }
